--- a/src/impl.cpp
+++ b/src/impl.cpp
@@ -338,9 +338,8 @@ void Manifold::Impl::CreateHalfedges(const Vec<ivec3>& triVerts) {
   // degenerate situations the triangulator can add the same internal edge in
   // two different faces, causing this edge to not be 2-manifold. These are
   // fixed by duplicating verts in SimplifyTopology.
-  stable_sort(ids.begin(), ids.end(), [&edge](const int& a, const int& b) {
-    return edge[a] < edge[b];
-  });
+  radix_sort_with_key(ids.begin(), ids.end(),
+                      [&edge](size_t i) { return edge[i]; });
 
   // Mark opposed triangles for removal - this may strand unreferenced verts
   // which are removed later by RemoveUnreferencedVerts() and Finish().