… extras/stl_test.cpp
… src/sparse.h