Closes #1176. This is a rather large change to the algorithm implementation, but it scales much better now.

```
                               1K    5K   10K   50K  100K    1M   10M 
sort_Rnd                     2.41  1.09  1.01  4.96  8.53 17.33 11.96 
sort_Rnd_size                1.08  1.08  1.04  8.76 15.27 28.58 28.37 
sort_Rnd_size_merge          2.24  1.07  1.02  4.81  9.15 15.66 10.51 
sort_roughly_Asc             1.25  1.03  0.99  1.40  2.30  2.73  2.64 
sort_roughly_Asc_size        1.00  1.06  1.02  1.09  1.62  6.52 12.00 
sort_roughly_Asc_size_merge  1.29  1.00  1.00  1.54  2.03  2.66  2.61 
sort_Des                     1.01  0.95  0.94  1.12  1.87  2.54  2.57 
sort_Des_size                1.02  0.98  0.99  0.86  1.34  4.33  5.64 
sort_Des_size_merge          1.01  0.94  0.96  2.13  2.73  3.28  2.95 
```

Speedup for radix sort for larger data size is increasing. Previously, it is not increasing due to merge being quite sequential:

```
                               1K   10K  100K    1M   10M 
sort_Rnd                     2.39  1.17 10.01 17.68 11.36 
sort_Rnd_size                1.67  2.86 12.66 22.35 11.90 
sort_Rnd_size_merge          2.42  1.04  9.25 16.13 10.71 
sort_roughly_Asc             1.31  1.02  2.13  3.14  2.69 
sort_roughly_Asc_size        1.13  1.50  3.16  4.86  7.29 
sort_roughly_Asc_size_merge  1.23  0.99  2.28  2.87  2.57 
sort_Des                     0.96  0.95  1.89  2.70  2.53 
sort_Des_size                1.26  1.37  2.05  2.63  2.94 
sort_Des_size_merge          0.93  0.91  2.36  3.03  2.82 
```

The new result is not as fast for smaller input instances, but that can probably be tuned.

On my computer, `Samples.CondensedMatter64` runtime reduced from 7785ms to 5365ms due to the huge sparse indices it has to sort. For other benchmarks, the improvement is not as significant (e.g. `Samples.Sponge4` from 2682ms to 2433ms). For perfTest, there is not much improvement, the collider is the slowest part there.