@pca006132 2025-03-06
Thinking about it, I think I have more things I can play with before deciding on the actual changes. LSB-radix sort isn't really that bad, in fact its sequential performance is great when the input is quite random, and maybe I can somehow change it to make it better for roughly sorted distributions.