@elalish 2025-03-07
I'm trying to read your perf tables - is the first column the name of the benchmark being tested? Are the values (time for old sort / time for this sort)? Also keep in mind that I'm planning to remove most of the SpareIndices usage - though we still sort in plenty of other places.