@pca006132 2025-03-07
Yes, but sort performance can depend a lot on the input distribution. I think we can withhold this for a while until we finish removing `SparseIndices` sort, and do optimization using the performance at that time.