@starseeker 2025-04-30
@pca006132 Is there an issue or PR for the "removing SparceIndices sort" goal?  Just thinking it would be handy to reference it here if there is...