@pca006132 2025-04-30
Oh sparse indices was removed entirely, so this is for everything else, e.g. halfedges.