@pca006132 2025-06-04
It seems that the performance of radix sort doesn't really matter now, we don't have that much things to sort, and bandwidth is the most limiting part. Changing sort implementation doesn't change that.